move level util to api definotions
