+++ src/views/battle/BattleResultComponent.tsx
@@ -13,7 +13,7 @@ import { wait } from '../../common/utils';
 import { CSS_COLOR } from '../../common/styles';
 import { ScenesService } from '../scenes/ScenesService';
 import { EffectsService } from '../../common/services/EffectsService';
-import { getLevelExperience } from '../../common/api/definitions/LevelUtil';
+import { getLevelExperience } from '../../common/api/definitions/levelUtil';
 
 export const BattleResultComponent = ({
     victoryResults,
